Melhorando a estrutura do testes para publico alvo invalido e cadastro de um curso salvo com mesmo nome
