+++ src/CurosOnline.Dominio/Curso/ArmazenadorDeCurso.cs
@@ -1,0 +1,32 @@
+using System;
+
+namespace CursoOnline.DominioTest.Cursos
+{
+    public class ArmazenadorDeCurso
+    {
+        private readonly ICursoRepositorio _cursoRepositorio;
+
+        public ArmazenadorDeCurso(ICursoRepositorio cursoRepositorio)
+        {
+            _cursoRepositorio = cursoRepositorio;
+        }
+
+        public void Armazenar(CursoDto cursoDto)
+        {
+            var cursoJaSalvo = _cursoRepositorio.ObterPeloNome(cursoDto.Nome);
+            
+            if(cursoJaSalvo != null)
+                throw new ArgumentException("Nome do curso já consta no banco de dados.");
+            
+            //Enum.TryParse<PublicoAlvo>(cursoDto.PublicoAlvo, out var publicoAlvo);
+
+            if (!Enum.TryParse<PublicoAlvo>(cursoDto.PublicoAlvo, out var publicoAlvo))
+                throw new ArgumentException("Publico Alvo invalido");
+            
+            var curso = 
+                new Curso(cursoDto.Nome, cursoDto.CargaHoraria, (PublicoAlvo)publicoAlvo, cursoDto.ValorDoCurso);
+            
+            _cursoRepositorio.Adicionar(curso);
+        }
+    }
+}
