+++ src/CurosOnline.Dominio/Curso/CursoDto.cs
@@ -1,0 +1,10 @@
+namespace CursoOnline.DominioTest.Cursos
+{
+    public class CursoDto
+    {
+        public string Nome { get; set; }
+        public double CargaHoraria { get; set; }
+        public string PublicoAlvo { get; set; }
+        public double ValorDoCurso { get; set; }
+    }
+}
