+++ src/CurosOnline.Dominio/Curso/ICursoRepositorio.cs
@@ -1,0 +1,8 @@
+namespace CursoOnline.DominioTest.Cursos
+{
+    public interface ICursoRepositorio
+    {
+        void Adicionar(Curso curso);
+        Curso ObterPeloNome(string nome);
+    }
+}
